// src/app/core/enums/modules.enum.ts
export enum Modules {
    Home = 'Home',
    DashboardCommercial = 'DashboardCommercial',
    DashboardProduction = 'DashboardProduction',
    Clients = 'Clients',
    Fournisseurs = 'Fournisseurs',
    Agences = 'Agences',
    Users = 'Users',
    Produits = 'Produits',
    Parameters = 'Parameters',
    Dossiers = 'Dossiers',
    AgendaCommercial = 'AgendaCommercial',
    Devis = 'Devis',
    Facture = 'Facture',
    Avoir = 'Avoir',
    Paiement = 'Paiement',
    Accounting = 'Accounting',
    CategoryProduct = 'CategoryProduct',
    ModesReglement = 'ModesReglement',
    BankAccount = 'BankAccount',
    AgendaParametrage = 'AgendaParametrage',
    ModeleSms = 'ModeleSms',
    TypeChauffage = 'TypeChauffage',
    TypeLogement = 'TypeLogement',
    ChampSiteInstallation = 'ChampsSiteInstallation',
    BonCommande = 'BonCommande'
}